💄 style    : move mypage components in app/mypage/components
Enhance page modularity and facilitate separate management from shared components
