+++ src/app/mypage/_components/Info/BasicInfoSection/basicInfoSection.style.ts
