+++ src/app/mypage/_components/Info/basicInfoConstants.ts
