+++ src/app/mypage/_components/Inquiry/InquiryDetail/InquiryDetail.tsx
