+++ src/app/mypage/_components/Inquiry/InquiryDetail/inquiryDetail.style.ts
