+++ src/app/mypage/_components/Inquiry/InquiryList/InquiryListItem.tsx
