+++ src/app/mypage/_components/Inquiry/InquiryList/inquiryListItem.style.ts
