+++ src/app/mypage/edit/page.tsx
@@ -2,12 +2,12 @@
 import { useState, useEffect } from 'react';
 import { useRouter } from 'next/navigation';
 import CommonButton from '@/components/common/Button/CommonButton/CommonButton';
-import ProfileSection from '@/components/Mypage/Info/ProfileInfoSection/ProfileInfoSection';
-import BasicInfoSection from '@/components/Mypage/Info/BasicInfoSection/BasicInfoSection';
+import ProfileSection from '../_components/Info/ProfileInfoSection/ProfileInfoSection';
+import BasicInfoSection from '../_components/Info/BasicInfoSection/BasicInfoSection';
 import EmailAuthModal from '@/components/Modal/EmailAuthModal/EmailAuthModal';
 import AddressModal from '@/components/Modal/AddressModal/AddressModal';
 import useModal from '@/hooks/useModal/usemodal';
-import { addActiveToBasicInfoData } from '@/components/Mypage/Info/basicInfoConstants';
+import { addActiveToBasicInfoData } from '../_components/Info/basicInfoConstants';
 import { UserInfo } from '@/types/apis/userInfo';
 import * as S from './edit.style';
 
@@ -15,7 +15,8 @@ export default function Page() {
   const router = useRouter();
   const { open, close, Modal } = useModal();
   const [userInfo, setUserInfo] = useState<UserInfo>({
-    userName: '',
+    userFirstName: '',
+    userLastName: '',
     userId: '',
     userEmail: '',
     address: '',
