+++ src/app/mypage/layout.tsx
@@ -1,4 +1,4 @@
-import SideNav from '@/components/Mypage/SideNav/Sidenav';
+import SideNav from './_components/SideNav/Sidenav';
 import * as S from '@/app/mypage/info.style';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
